backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
openai==1.37.0
redis==5.0.7
sqlmodel==0.0.16

